niorlusx_call_service/requirements.txt
Flask
openai
python-dotenv
twilio
